promote cacheid type to long int and fix update of the stage counter (#1953)

+++ libs/common/procedural_reader.cpp
@@ -79,7 +79,7 @@ void ProceduralReader::Read(const std::string &filename,
     _overrides = nullptr; // clear the overrides pointer. Note that we don't own this array
 }
 
-bool ProceduralReader::Read(int cacheId, const std::string &path)
+bool ProceduralReader::Read(long int cacheId, const std::string &path)
 {
     if (!GetNodes().empty()) {
         return true;
